Fix config cast type

+++ data_syncer/config.py
@@ -30,7 +30,7 @@ ASYNC_METRICS = config("ASYNC_METRICS", default=True, cast=bool)
 STORE_METRICS_LOCALY = config("STORE_METRICS_LOCALY", default=False, cast=bool)
 
 # The interval between each temp update from the firing schedule program
-PROGRAM_UPDATE_INTERVAL = config("PROGRAM_UPDATE_INTERVAL", default=60, cast=int)
+PROGRAM_UPDATE_INTERVAL = config("PROGRAM_UPDATE_INTERVAL", default=60, cast=float)
 
 # The interval between each state sync with the arduino board
 LOOP_UPDATE_INTERVAL = config("LOOP_UPDATE_INTERVAL", default=1, cast=float)
